temporal_baseline/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy
opencv-python
Pillow
scikit-image
tqdm
lpips
